move web utilities to cli/js/web/util.ts, remove types.ts
